fix ui  performance
1. label ui drawcall too much high
2. switch visibility when parent changed
