+++ cocos/3d/ui/components/ui-component.ts
@@ -1,6 +1,9 @@
 import { Component } from '../../../components';
+import { SystemEventType } from '../../../core';
 import { ccclass, executeInEditMode, executionOrder, property } from '../../../core/data/class-decorator';
 import { UI } from '../../../renderer/ui/ui';
+import { Node } from '../../../scene-graph';
+import { CanvasComponent } from './canvas-component';
 
 /**
  * @zh
@@ -41,9 +44,50 @@ export class UIComponent extends Component {
 
     protected _visibility = -1;
 
+    private _lastParent: Node | null = null;
+    public onEnable () {
+        this._lastParent = this.node.parent;
+        if (this._lastParent) {
+            this.node.on(SystemEventType.CHILD_REMOVED, this._parentChanged, this);
+        }
+    }
+
+    public onDisable () {
+        this._cancelEventFromParent();
+    }
+
     public updateAssembler (render: UI) {
     }
 
     public postUpdateAssembler (render: UI) {
+    }
+
+    private _parentChanged (node: Node){
+        if (node === this.node) {
+            let parent = this.node;
+            // 获取被渲染相机的 visibility
+            while (parent) {
+                if (parent) {
+                    const canvasComp = parent.getComponent(CanvasComponent);
+                    if (canvasComp) {
+                        this._visibility = canvasComp.visibility;
+                        break;
+                    }
+                }
+
+                // @ts-ignore
+                parent = parent.parent;
+            }
+            this._cancelEventFromParent();
+            this._lastParent = this.node.parent;
+        }
+    }
+
+    private _cancelEventFromParent (){
+        if (this._lastParent) {
+            this._lastParent.off(SystemEventType.CHILD_REMOVED, this._parentChanged, this);
+            this._lastParent = null;
+            this._visibility = -1;
+        }
     }
 }
